+++ ExpressionUtil/ICountable.cs
@@ -1,0 +1,5 @@
+﻿namespace NINA.InstructionMath.ExpressionUtil {
+    public interface ICountable {
+        int Count { get; }
+    }
+}