+++ CHANGELOG.md
@@ -2,3 +2,13 @@
 
 ## 1.0.0.1
 - Initial release
+- Added Math Loop operator
+  - Allows users to specify a math expression which determines when the loop ends
+
+- Added variables
+  - count - the number of times the loop has completed
+  - time - The current time (ms since epoch)
+  - astro_dawn - The time of the next astronomical dawn (ms since epoch)
+  - astro_dusk - The time of the next astronomical dusk (ms since epoch)
+
+- Added expression editor to assist in writing the math expression